Reset AR model cache between tests
When models are created fresh for each test the configuration may be cached by
ActiveSupport. This results in tests failing if they are executed after others
that use the same model names with different associations. The method for
caching this information was changed for Rails 7 meaning that is is no longer
necessary to explicitly clear the cache. The change was made here:

  https://github.com/rails/rails/commit/14d4edd7c3b06e82e1fcef54fa0b4453315c35fd